use sqrt default type

+++ tests/test_weighted_mean_estimator.py
@@ -76,9 +76,7 @@ def weights(sim, weights_dtype):
     # Construct simple test weights;
     # one uniform positive and negative weighted volume respectively.
     r = 2  # Number of weighted volumes
-    weights = np.full(
-        (sim.n, r), 1 / np.sqrt(sim.n, dtype=weights_dtype), dtype=weights_dtype
-    )
+    weights = np.full((sim.n, r), 1 / np.sqrt(sim.n), dtype=weights_dtype)
     weights[:, 1] *= -1  # negate second weight vector
 
     return weights
